Clear up output of code blocks
The first two code blocks that involve an instance variable called `@transactions` indicate that the output would be `#=> ["10.00", "-15.00", "25.00", "30.00", "-24.00", "-70.00", "999.00"]`, when in reality, if you run the code, the output ends up being each of those values separately as strings. This can be confusing to readers, since with the given output, it is implied that the method pushes the values into an array at some point.